A couple of UI brushups.
Tooltips are shown to the right. When they are shown to the top, they
don't always have enough time to fade out when moving to the ip above.
This causes that the tooltip for the ip above isn't shown sometimes.
Another workaround is to switch off fading.
Another fix is adding min-width to column with hacking control buttons.